+++ lib/searchkick.rb
@@ -9,6 +9,7 @@ require "searchkick/reindex_queue"
 require "searchkick/hash_wrapper"
 require "searchkick/results"
 require "searchkick/query"
+require "searchkick/multi_search"
 require "searchkick/model"
 require "searchkick/tasks"
 require "searchkick/middleware"
@@ -37,7 +38,7 @@ module Searchkick
   class ImportError < Error; end
 
   class << self
-    attr_accessor :search_method_name, :wordnet_path, :timeout, :models, :client_options, :redis, :index_suffix, :queue_name
+    attr_accessor :search_method_name, :wordnet_path, :timeout, :models, :client_options, :redis, :index_prefix, :index_suffix, :queue_name
     attr_writer :client, :env, :search_timeout
     attr_reader :aws_credentials
   end
@@ -102,14 +103,8 @@ module Searchkick
     end
   end
 
-  def self.multi_search(queries)
-    if queries.any?
-      responses = client.msearch(body: queries.flat_map { |q| [q.params.except(:body), q.body] })["responses"]
-      queries.each_with_index do |query, i|
-        query.handle_response(responses[i])
-      end
-    end
-    queries
+  def self.multi_search(queries, retry_misspellings: false)
+    Searchkick::MultiSearch.new(queries, retry_misspellings: retry_misspellings).perform
   end
 
   # callbacks
@@ -154,6 +149,8 @@ module Searchkick
         completed: batches_left == 0,
         batches_left: batches_left
       }
+    else
+      raise Searchkick::Error, "Redis not configured"
     end
   end