+++ lib/searchkick/index_options.rb
@@ -162,11 +162,6 @@ module Searchkick
           settings[:similarity] = {default: {type: options[:similarity]}}
         end
 
-        unless below60
-          settings[:mapping] ||= {}
-          settings[:mapping][:single_type] = false
-        end
-
         settings.deep_merge!(options[:settings] || {})
 
         # synonyms
@@ -188,7 +183,7 @@ module Searchkick
           # - Only apply the synonym expansion at index time
           # - Don't have the synonym filter applied search
           # - Use directional synonyms where appropriate. You want to make sure that you're not injecting terms that are too general.
-          settings[:analysis][:analyzer][default_analyzer][:filter].insert(4, "searchkick_synonym")
+          settings[:analysis][:analyzer][default_analyzer][:filter].insert(4, "searchkick_synonym") if below60
           settings[:analysis][:analyzer][default_analyzer][:filter] << "searchkick_synonym"
 
           %w(word_start word_middle word_end).each do |type|