+++ CHANGELOG.md
@@ -1,6 +1,14 @@
-## 2.3.2 [unreleased]
+## 2.3.2
 
 - Added `_all` and `default_fields` options
+- Added global `index_prefix` option
+- Added `wait` option to async reindex
+- Added `model_includes` option
+- Added `missing` option for `boost_by`
+- Raise error for `reindex_status` when Redis not configured
+- Warn when incompatible options used with `body` option
+- Fixed bug where `routing` and `type` options were silently ignored with `body` option
+- Fixed `reindex(async: true)` for non-numeric primary keys in Postgres
 
 ## 2.3.1